Fix: ID column sorting using numeric comparison instead of string
- Cast event.getId() to long long in GetSortValue() to match the numeric type check in sort()
- Ensures ID column sorts numerically (0, 1, 2, ...) instead of lexicographically (0, 1, 10, 2, ...)

+++ src/application/ui/qt/EventsTableModel.cpp
@@ -456,7 +456,7 @@ QVariant EventsTableModel::GetSortValue(const db::LogEvent& event,
     const std::string& columnName) const
 {
     if (columnName == "id")
-        return QVariant::fromValue(event.getId());
+        return QVariant::fromValue(static_cast<long long>(event.getId()));
     
     if (columnName == "source")
         return QString::fromStdString(event.GetSource());
